Added functions for detecting alphabet in MEME format. In addition to now using explicit definitions with ALPBABET=DNA/RNA/AA, read_meme can now detect DNA/RNA/AA-LIKE alphabets, and custom alphabets (note: custom alphabet functionality is likely to cause downstream failures, see get_custom_meme_alph for details.
This change required also changing how alph.len was calculated. Now instead of
using nchar() from explicit definition, uses lookup table for default types, or
counts nchar() for custom alphabets (this is the likely point of failure for
custom alphabets, as alph.len for a lookup table will be longer than the number
of letters used in the matrix). I think this is acceptable for now, because
custom alphabets currently always fail on import, so this shouldn't cause
regressions.

Added `dreme.txt` to `inst/extdata/` and a unit test to test_read_motifs.R to test this new funcitonality.

test_read_motifs.R passes all tests with this change.