Improves support for MEME custom alphabet. Closes #24.
This change adds some error handling & additional catches to support
weirder versions of custom alphabet implementations in the meme suite
output files. Adds a unit test to check for functionality. NOTE: this
implementation is a bit hacky, but I cannot find a spec for how custom
alphabets are indicated in MEME output files, so this reactive-type
approach is all I could come up with.